Isolated active collection behaviours into own trait.
