+++ test/core/document/test_validation.py
@@ -2,11 +2,11 @@
 
 from __future__ import unicode_literals
 
-from marrow.mongo import Document
 from marrow.mongo.field import Array, String
+from marrow.mongo.trait import Collection
 
 
-class StringDocument(Document):
+class StringDocument(Collection):
 	optional = String()
 	nonoptional = String(required=True)
 	choose = String(choices=["Hello", "World"])
@@ -21,8 +21,3 @@ def test_string_document():
 				'nonoptional': {'$type': 'string'},
 				'choose':  {'$or': [{'$exists': 0}, {'$type': 'string', '$in': ['Hello', 'World']}]}
 			}
-
-
-
-class ArrayDocument(Document):
-	values = Array(String())
